Board class header file created
Board class is defined with its data members. It is going to have `Tile` as its data member which implies it is Composition relationship between Board and Tile classes. Board is composed of Tiles we can say. Methods for Board.cpp are being written. It is a Work In Progress.